Added light mode
Added reference to lightstyle.css
Added light mode button that switches to dark mode text when light mode is enabled using Javascript
Javascript to the end of document, that listens for an onclick event of a navigation button. When button is pressed a Javascript functions runs the .light-mode CSS style on the document page
To keep the light mode persistent, a local session storage is used to store if the user choice is light mode, and runs as the user browses to website.