Assignment 2 part 3
removed logo, navigation and footer html content that is used on all pages on this website in exchange for javascript dynamically inserting it from a .js file. 

This avoid the same code being written on each page of this website that uses it, reducing size and clutter of the entire code.