Assignment 2 part 2 - Blog
Added a dynamic blog that retrieves blog posts from a json file and populates a list of blogs. The user can select a blog, and based on its id, javascript will load the blog's content, removing html elements withing the body and inserting a new article element with the content of the selected blog post. The script for this will be in an external file scripts.js